requirements.txt
robotframework>=4.1.3
robotframework-appiumlibrary>=2.0.0
Appium-Python-Client>=2.6.1
urllib3==1.26.6
requests>=2.28.1
python-dotenv>=0.19.2
allure-robotframework==2.13.2
robotframework-pabot==2.16.0
robotframework-seleniumlibrary==6.1.0
# robotframework-process>=1.0.0  # Geçici olarak kaldırabilirsiniz 